SharepointShareplumDemo/requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
cycler==0.11.0
DateTime==4.4
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.2
lxml==4.8.0
matplotlib==3.5.2
ntlm-auth==1.5.0
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
SharePlum==0.5.1
six==1.16.0
urllib3==1.26.9
wincertstore==0.2
zope.interface==5.4.0
